v0.15.0 (#225)
* Fix slash event on double sign (full resync needed)

* fix version

* Fix double sign slashing issue #215

* skip already offline candidates to prevent double punishing

* Make block hash lowercase

* Remove unused fields from config

* Fix issue #209

* Update changelog

* Delete coins with 0 reserves #217

* Add test

* Export/import app state

* Import state

* Fix genesis

* Add tests

* Switch to Amino and fixes

* Fix frozen funds height

* Refactor

* Fix

* Add ?include_stakes to /candidates endpoint #222

* Update tendermint to v0.31.0

* Update tendermint to v0.31.0

* UpdaUpdate tendermint to v0.31.0

* Add CoinLiquidation event #221
Set height to be always in uint64

* Fixes

* Change `stake` to `value` in DelegateTx

* Invariants

* Fix nil pointer exception

* Fix nil pointer exception

* Record total slashed in state

* Fix record total slashed in state

* Fix coin creation issue

* Fix genesis

* Fix node termination

* Fix invariants

* Fix mempool issue #220

* Fix tests

* Recheck mempool once per minute

* Small refactor

* Change `pubkey` to `pub_key` in all API resources

* Add invariants checker each 720 blocks

* Fix events issue

* Fix invariants

* Fix gui

* Fix gui

* Fix invariants

* Create new state for invariants checking

* Fix coin creation issues

* Compute validator power from genesis

* Update genesis

* Update config

+++ api/transaction.go
@@ -13,7 +13,7 @@ func Transaction(hash []byte) (*TransactionResponse, error) {
 		return nil, err
 	}
 
-	if tx.Height > blockchain.LastCommittedHeight() {
+	if uint64(tx.Height) > blockchain.LastCommittedHeight() {
 		return nil, rpctypes.RPCError{Code: 404, Message: "Tx not found"}
 	}
